Lex contextual keywords
This now lexes context-specific keywords into unique tokens. This will
make it easier to parse the various named tables.

A consequence of this is that code that was previously expecting idents
may fail if the ident is a contextual keyword. To get around this, we
can now pass `TokenSet` in most places where we previously used kind
(such as Parser::eat) and we have a TokenSet::IDENT_LIKE const that
includes all of the contextual keywords.