Track absolute position of nodes&tokens
This was a long time coming, and makes life easier in a lot of ways.
However, the implementation is a bit weird: the position is tracked
in the cursor, and updated as the cursor moves. This means that if
you were to access nodes/tokens directly, before a traversal, they
may not have the correct ranges.

To make this less likely, I've removed all the places where children
are accessed directly; you now need to go through a cursor.