Ability to save known-okay ttx diffs as passes
If there is a XX.expected_diff file in the fonttools-tests file,
and if the diff of our ttx and the fonttools ttx exactly matches the
contents of this file, we will consider it a pass.

These files can be generated by the -d/--write-diff flag on the
ttx_test cli utility.