Rewrite contextual sub nodes after parsing.
This was an alternative to the lookahead parsing business. I thought the
tradeoffs might be better, here, but they seem to be just as bad.

My thinking was: if we do it this way, we have the advantage of not
compromising on parser speed, and then also that we could reuse things
like glyph classes that we've already parsed once.

This didn't work as well in practice. In particular, all of our parsing
code works on tokens through `Parser`, and so we would need to rewrite
tons of stuff. More generally, though, trying to work with nodes during
parsing is something we had not considered when designing various other
parts of the API, and it's a weird fit.

This new strategy seems like a better compromise: for nodes that need to
be reparsed, we assign some temporary Kind value (such as
Kind::GsubNodeNeedsRewrite) and then whenever a node is finished, we
have an opportunity to rewrite the children and pick a new Kind
beforehand.