Tweak gsub parsing and validation
This adds clearer reporting of when a test failure is an 'unexpected
success', tweaks parsing chain rules, and we now handle 'lookup and inline
rule in same statement' during validation, not parsing.

This is all sort of arbitrary, trying to split the difference between
providing good inputs to later stages and also providing good diagnostic
messages.