Kind of hacky way of breaking up anonymous tables
When handling contextual rules, you cannot always put all inline
rules in the same table, because they might collide. For instance given
the two rules,

pos B B' +10;
pos A B' +20;

we have two different single pos rules for glyph 'B'; these can't go in
the same lookup.

It feels like there is lots of room for optimization, here; for instance
we could just put the rule in the first lookup of the appropriate type
for which there was no existing rule. For now, though, we are happy to
just create new lookups as needed; but we need to figure out when these
collisions happen.