More robust lookup splitting
During initial development, I didn't know if I would be expecting
features to have mixed GPOS/GSUB lookups, and my initial logic for
dealing with them, if they occured, was naive to the point of useless.

This reworks that logic to be much more robust, splitting lookups
manually if needed.

This also includes a new test case in the test-data submodule.