Support in-memory compilation
This patch introduces a new trait, `SourceResolver`, which abstracts out
the task of finding and loading the FEA for a given path, both for the root
source as well as for any additional sources referenced by `include($source)`
statements.

The logic around includes in FEA is complicated; depending on the
implementation, include statements may be resolved relative to the
*root* fea file, to the *including* file, or to a parent directory.
Previously this resolution logic was spread out across various places;
it now is generally isolated to the internal `FileSystemResolver` type.