add pre-commit git hook to check for println! or eprintln! statements in Rust files
It only checks the lines that are actually added (cached) to the index, not all the lines in the .rs file. It also tries to be clever strip all `//` inline comments so that a commented-out println!() is allowed.
If we wanted to support multi-line /**/ comments, there's a crate/CLI tool called comment-strip but I don't think it's worth it.