[fontbe/head] set created/modified to current timestamp or SOURCE_DATE_EPOCH
This matches fontmake behaviour of setting the head.created/modified timestamp to the current time (in seconds since the TrueType epoch), or to SOURCE_DATE_EPOCH if the environment variable is present (see: https://reproducible-builds.org/specs/source-date-epoch/).

The head.created can also be controlled via 'date' font attribute in .glyphs source or via 'openTypeHeadCreated' in UFO fontinfo.plist, but we can implement that later (I'll file an issue for that).

This is sufficient to at least prevent fontTools to issue a warning when the created/modified fields are set to 0 (because they look suspiciously old...).