Always check for expected diff in ttx tests
In our reuse of the fonttools test suite, we have a mechanism whereby we
can indicate that certain diffs in our ttx are acceptable; for instance
in some places we may choose not to implement a given feature, to choose
a different table format, or otherwise produce a binary that is
acceptable but is not *exactly* what fonttools produces.

Previously, if there was an expected diff for a given input, but we
produced output identical to fonttools, we would succeed. This means
that we had a bunch of expected diffs sitting around that were never
actually used.

With this change, we will now fail a given test case if there is an
expected diff, but our output is identical to fonttools.

In addition, we now ignore comments when comparing these diffs; I would
like to adopt the convention that the expected diff files include an
explanation of why we believe our output to be acceptable.