Use stable ConditionSet sort order
I don't believe this is mentioned anywhere in the spec, but it is an
important detail: the sort order of conditionsets in the
FeatureVariations table has meaningful impact on whether a variation is
applied, since we apply the first set of variations that is encountered.

With this patch, we match the order used by fonttools, which is (as far
as I can deduce) based on the order in which a given conditionset is
first referenced by a lookup.