Fixup loading of test-data for non-tests
I call a function that loads a file from a relative path. It is called
two different ways; either when running `cargo test`, or when running
the ttx_test binary.

`cargo test` runs tests for a given crate from that crate's root
directory, but cargo run --bin=x can be run from the crate root; so I
need to ensure that I use the correct relative for both of these cases.

I really don't know what the principled approach to this is, outside of
baking the file into the binary at compile time? But this at least fixes
my concrete problem.