Read with timeout instead of blocking
Previously we were using a blocking read to wait for completions, which
would deadlock if a worker thread had panicked.

Now we will read in a loop, and return an error if we timeout.

I've arbitrarily decided that ten seconds is a reasoanble timeout
interval, but it might not be! We also may wish to make this setting
configurable.