Halt runloop when jobs panic
This involved quite a bit of fiddling. I couldn't get rayon's built-in
panic_handler to work, so I ended up rolling my own solution, where we
manually catch unwinds that occur in spawned jobs.

When a panic occurs, we will set an atomic flag and return an error. We
check this flag before beginning any scheduled job, and return
immediately if it is set.