More careful zeroing of value records
In general when a value record contains a 0 value for an advance or
placement, we want to treat this as "no value", since this lets us more
efficiently pack the data.

In the case of a variable font, however, it is entirely possible for a
value record to have a default value of 0. In this case we want to leave
this field in the record.

I'm not 100% sure this makes sense, since if the implementation is variable-
aware it can compute the correct value, and if it isn't it shouldn't matter?
But erring on the side of "do what fonttools does, for now."