[ufo2fontir] Fix for optional designspace names
The upstream API has changed to make the 'name' fields on
instance/source types optional. This patch is the minimal fix, where we
check all names on load and insert placeholders if they are missing.

An alternative solution might be to use something other than the name to
uniquely identify these objects.