VarStore: skip writing deltas for zero regions
So I see that there was a bug in my logic, and I know how to fix it, but
I do *not* know what conditions were triggering it that are not covered
in the tests, and I would like to.

That said, this is a fix that works and should unblock downstream work,
so I  think it's worth merging as-is; I'll look into adding a test case
shortly.