[ci] Remove code to ignore certain test cases in CI
This was gross; the idea was that we had certain known failures in the
fonttools test suite, but we wanted to run the rest of those cases for
the time being, so would explicitly check to see if we were in CI when
running tests.

This was probably too cute. In any case there are now PRs up that
resolve these outstanding failures, so we can get rid of this.