factor out code to convert fontir::VariationRegion => write-fonts' equivalent
it's used in a couple of places already. It does not need to be fallible, in the unlikely scenario a tent for a given axis is missing from a region we can assume the no-op (0,0,0). The VariationModel and the static_metadata.axes should already be in sync, but even if the region contains less axis coordinates than there are fvar axes, it's clear what the omission means ('axis does not participate')