[glyphs2fontir] don't expect all glyphs to define min and max, default is enough
the whole point of supporting Virtual Masters is that not all glyphs need to define the same set of locations.

Also, there was a typo.. the code was confusing 'max' with 'default'

Removed MinUndef.glyphs for the check no longer applies. In fact, if you try open that file in Glyphs.app the master layer that had been deleted probably using a text editor reappears, because the expectation is that all glyphs have exactly one layer per (non-virtual) master.