don't emit IR by default and rename option to --incremental
As discussed with Rod in chat, we prefer to have incremental builds be opt-in rather than opt-out. Not only it's faster to build without, but also it reduces the risk that the casual user might incur in potential bugs on subsequent builds related to incomplete inter-task dependencies.

Also I propose to rename the option from --emit-ir to --incremental because it communicates the intent better to the user. The old name is kept for backward compatibility but it won't be shown in the --help.

The flag takes an optional [true|false] argument, which can be omitted altogether. E.g. --incremental is equivalent to --emit-ir=true; omitting the flag by default is equivalent to --emit-ir=false or --incremental=false.