[fontbe/avar] don't skip no-op segment maps
GS contains a GRAD axis whose avar segment maps does not have any interesting mappings. It happens to be the last one in the axis order. We were filtering out axis segment maps that contain only identity maps, leading to an avar table with fewer segment maps than there are fvar axes, which is invalid as OTS correctly points out. So we now still emit these when bulding avar and there's at least one segment map that is not an identity map.
Note that, despite the OT spec allows for empty segment maps for axis that don't distort the default normalization, we want to match fontmake/fonttools and emit the required {-1:1, 0:0, 1:1} mappings even when no other interesting mappings are present for an axis.

Fixes #582 avar axis mismatch OTS error

after this GS avar table is the same when built with fontc vs fontmake.