[fea-rs] Disable printing in test
I believe this was turned on by accident when I migrated fea-rs into
this repo, and I noticed it was sometimes showing up in test output.

This was confusing, since normally rustc swallows stderr during test
execution, and only prints it if a test fails? but apparently this
depends on some implementation detail of `eprintln`, and does not apply
when you do the `writeln` hack.