[fontir] Remove custom serde impl for Kerning
The problem was that serde_yaml doesn't like it when BTreeMap is used as
the key to a collection. Since this is only the case in `Location`, it's
much simpler to just have a cusom serde impl for that type.

I expect this will lead to further simplification, although I will leave
that for additional PRs.