[fontir] Simplify StateSet serialization
The problem here is that serde really does not want to use an enum as a
key. Fortunately for us, we don't really need the key to be an enum,
since the value in the map encodes whether or not something is a path or
not. This means we can simplify things and just use PathBuf as the key
type, and treat this purely as an implementation detail.