[fontc] Validate glyph fliter regex at parse time
This removes the generally unhelpful 'BadRegex' error variant and
ensures that if a glyph filter is passed at the command line, we ensure
that it is valid regex, and let clap handle displaying a helpful error
if not.

This requires adding a new type, which is needed because Regex does not
implement the serde traits or PartialEq.