[ttx_diff] Better handling of pairpos subtables
This changes the logic around skipping rules to only skip if there is an
entry for the _pair_, instead of for the first glyph.

Most of this PR is a test and a bunch of test-setup code which I hope to
reuse.

This uses the builders in fea-rs, and doing that required me to make a
trait in there public, but that seems fine