[fea-rs] Fixups for ttx_test util
This utility had broken a bit during the move from the old fea-rs repo
into this new repo; in particular it would fail if you ran it from the
`fea-rs` directory, which is annoying.

We will now attempt to figure out where we're running from when
determining the path to the test data, instead of hardcoding paths.