[fea-rs] Update fontools test cases
This mostly involved deleting a bunch of '.expecte_diff' files we had
that represented various small bugs & inconsistencies in fonttools.
All of those issues have now been resolved, and our output now matches
fonttools more closely.

This does involve one 'regression', which is that fontools has
implemented an additional optimization that we don't have. I'll open an
issue for this separately, and in the meantime I've added a
.expected_diff for this case (bug512.fea)

I've also had to disable GSUB_2.fea, because it was modified to include
fonttools-specific syntax that fea-rs does not yet support.