Add KernGroup type
This encodes in the type system the idea that kerning groups have an
associated side, instead of having this be encoded as a prefix on a
string.

This removes the need to hae separate `KernSide` enums, and removes
possible complications around using the same type for glyph names and
group names.