[kerning] Split kern rules by script & directionality
This is very closely based on the code in the KernFeatureWriter, in
ufo2ft. With this patch we very nearly match fontmake's output for
kerning in oswald, with a few lingering differences.

It feels like there is room for polish here, but I also think it's worth
checkpointing here, for further iteration.