Guard against empty parent path
I was seeing a panic where we were trying to create the directory "" if
I passed `fontc -o myfont.ttf` e.g. a relative path to a file. Per the
docs on Path::parent,

> .. Returns `Some("")` for relative paths with one component

So we need to explicitly check this case.

In the process I also tweaked the signature of require_path; there's no
need to return a `PathBuf` that is identical to the path that was passed
by the user (and we were always throwing away the returned path in any
case.)