[fontbe] Fixup errors for variable metrics
There were a few giant FIXMEs in the resolve_variable_metric fn; if
memory serves I think I'd initially written that patch as "just a
sketch" and then it was merged to unblock other work.

In any case, this rationalizes things a bit, adding error variants for
the case where we can't compute deltas for anchors & kerns.