[g2f] Use smolstr when parsing glyphs.app data
Particularly for types that end up being SmolStr later on, like glyph
names; it's inconvenient and inefficient to parse to String and then
later convert to SmolStr, so this is a minor simplification.

Arguably the whole plist parsing process should be using smolstr, which
should offer a reasonable speedup, but we can worry about that when
we're thinking more generally about optimization.