Add fontdrasil::glyph_data module and codegen
This is based on the non_spacing_marks.py script that is being used in
glyphs2fontir.

In order to properly generate mark positioning lookups, I need access to
the categories and subcategories that glyphs.app assigns to glyphs. This
patch adds a new module to fontdrasil that exposes this data. Currently
this is limited to mapping from glyph names to category information, but
it would be trivial to update this to also provide conversion from
unicode values to glyph names (and thus from unicode values to
categories).