[fontc] Simplify timing color code
I was in here to add a new IR type and thought this could be cleaned up.

In addition to the cleanup, a tiny additional benefit of this change is
that gray is now the default, and you don't need to add an explicit case
anytime a new workid is added.