[glyphs] Store codepoints in glyph struct
Previously we were generating a separate cmap-like thing in the glyphs
module, which was then used to set the codepoints in the ir::Glyph. This
just keeps the codepoints in the glyphs Glyph, which will be nice
because it means that all the things that we use to determine the glyph
category will be in one place.