[glyphs] More and better tests for anchor propagation
This brings in two hand-crafted test files from cosimo, that cover many
of the cases exercised by the anchor propagation code.

This revealed a bunch of problems with my implementation. I've added a
separate test case for each failure I hit, but I'm also including a
blanket test on the raw sources as cover.