Update env-logger
The API changed a bit, so needed to fix that up. Also noticed that our
logs were never actually printing a thread name, and this doesn't seem
super useful given that the threading isn't something we manage, so I've
just removed that bit of the log preamble for the time being.