[fea-rs] Add test case for adjusting aalt lookup ids
When we generate the aalt feature we need to insert new lookups into the
beginning of the lookup list. This means we need to update all
subsequent lookup ids, and I was concerned that we were not doing this
correctly for lookups that were referenced by contextual rules.

It turns out we were handling this correctly, but it is a situation that
deserves a test case.