[fea-rs] Stop sorting lookups in feature
The order of lookups in the feature doesn't matter, so this is not a
functional change; prevoiusly we sorted because sorting is required for
the 'dedup' method on Vec to do what we want.

The sorting was something we did that fonttools didn't, and I'd like to
try and converge as much as possible.