[crater] Use rayon
In the case of ttx_diff in particular much of the work happens serially,
so running multiple fonts in parallel should be an easy win.

In some very simple benchmarking this reduced runtime by about 25%, but
that is with us failing to compile the majority of fonts and so really
not taking much advantage of the possible parallelization.