[fea-rs] Improve parsing of glyphsapp number values
This syntax is ambiguious, in that something like 'pad-2' is a valid
identifier per our normal parsing rules, but needs to be parsed as
(pad)(-)(2) for the purposes of the glyphs syntax.

This requires a adding a new method to our parser type that can split up
the current token.