[ttx_diff] Run otl-normalizer in parallel
For very large fonts one of the biggest time sinks is actually
normalization. We need to normalize both outputs; previously we were
doing this sequentially, which is totally unnecessary; with this patch
we ensure that we normalize both fonts at the same time (using python
threading, which is adequate since all the actual work is happening in a
child process.)

For a simple test input this reduced runtime from 12m28s to 8m56s.