[crater] Add basic 'detail view' for previous run
This now renders a very basic summary of changes and results for the
last run after the overall summary table.

In the future I would like to make it so this is hidden by default and
can be displayed as desired for the last N runs, but this first pass is
a useful start.

In addition I would like to add another level of detail for individual
runs (showing the command that failed and the backtrace) but that can be
followup.