[fontbe] Create GDEF for GlyphClassDef if needed
Previously we would only add a GlyphClassDef table to an existing GDEF
table, but would not create a GDEF table if one had not been produced by
feature compilation.

This patch matches the behaviour of fontmake.

This does raise the question of whether or not either compiler should
bother generating a GlyphClassDef subtable in the absense of GPOS/GSUB,
but that's something we can address separately.