[crater] Tidy ci report html
Previously we would just write a big contatenated string of html, which
was unreadable and also impossible to diff.

With this patch we clean up the html (newlines & indentation) using
tidy.

This is particularly important because it means git can efficiently
compute deltas on the index, instead of each new commit involving adding
a new ~2MB blob to the repo.