[normalizer] Consider lookup flags when merging rules
This was causing normalizer to be non-deterministic in some cases;
because we were merging rules without considering the lookup flags, we
would preserve the flag of the first rule we encountered, which would
depend on the iteration order of the hashmap.

This fixes the underlying issue, by not merging any rules that do not
have identical lookup flags.