[fea-rs] Fix include path lexing
statements such as `include(something.fea)` require special treatment,
because the stuff in the braces are irregular, and can contain arbitrary
symbols. To handle this we attempt to keep track of whether or not we
are currently lexing an include statement, but this logic was broken if
there was whitespace between the 'include' keyword and the parens.

In addition, we were not correctly handling leading whitespace _within_
the parens; to handle this (along with trailing whitespace) we now lex
everything in the parens as a path, and then split the token up if
necessary during the parsing pass.