[crater] Remove 'fonts_repo' cli arg
This has nothing to do with CI, and is only used by the old-style
runners.

Ultimately we should just unify these different interfaces; by default
we could keep a list of inputs in the cache dir as a json file, and with
no provided args we would create or run that.