[ttx_diff] Simplify reuse logic
At the beginning of the run, we now check the 'rebuild' flag, and
explicitly delete anything we are required to rebuild. This lets us
simplify later logic by knowing we're allowed to reuse anything that
exists.

This is part of a larger refactor to support the CI use case; because CI
builds each individual font in a transient temporary directory, these
artifacts do not persist between runs. To reuse artifacts CI is going to
need to start storing these artifacts somewhere other than in the build
directory.