[crater] Prebuild fontc & normalizer
Currently each invocation of ttx_diff attempts to `cargo build` these
targets. This is a problem if you want to run crater locally and also
continue working on fontc, since ttx_diff will attempt to recompile any
saved changes.

It is also a possible source of various intermittent 'other' errors
including some very clear instances where the binaries are not in the
target dir as expected.