[crater] Tweak how we save target paths
Instead of these being full paths, we save the path of the sources
directory and then the config and source files are paths relative to
that.

This mostly just lets us be represented more efficiently in json.

- This also makes config & source on target private, so the only way to
get these paths is to pass in the `cache_dir` and receive the fully
resolved path.