Remove BeValue type
This apparently only existed to get around the orphan rules for
`Persistable`, but we can just do that by moving the implementation into
the same crate that defines that trait (`fontir`).

In the process of this work I discovered that avar needs special
handling, because it is the one in this context that is allowed to
explicitly be `None`.

This had previously been handled by `BeValue`; to support this use case
I also added an explicit `PersistableOption` type, but thinking about it
more now I think this is unnecesary? Shouldn't it be fine to just
serialize an empty avar? But my quick attempt to do this ran into some
confusing test failures, so this seems fine for now.