[glyphs] Parse kerning as floats
Glyphs kern values can be floats, but we were parsing them as integers.

This also uncovered something else fishy: our integer parsing logic in
glyps-reader/ascii-plist-derive will happy parse something like "1.321"
as an integer, by parsing to a float and casting.

I think this is incorrect? I think it is reasonable to go the other way
(to accept an integer value as a float) and it _might_ be reasonable to
accept an integer value with no fractional component as a float, but I
definitely don't think we should parse a float with a fractional value
as an integer; it hides bugs like this.