We want more (f64)
This replaces a bunch of use of f32 (in parsing and in coordinate types)
with f64.

There are two tests that are failing which I don't want to touch.

This doesn't have a dramatic impact on crater, but it does show
improvements for fvar & HVAR for a number of fonts.