[glyphs] Impl 'erase open corners'
This implements the functionality to erase 'open corners', which are
features of some outlines where the intended corner of two curves needs
to be inferred.

This implementation is a bit more verbose than we'd like, because kurbo
does not currently have an implementation of arbitrary curve-curve
intersection, so we've ported the implementation from fonttools' bezierTools.