[ttx_diff] Mark adjusted points as 'diff_adjusted'
Previously it was just 'adjusted', which is general enough that it was
unclear to me when I saw it in diffs that it had originated in the
ttx_diff script, and not in the original ttx output.

This name will hopefully make it more clear that this is a diff-specific
thing.