[glyphs] Don't rotate nodes if all are off-curve
Although it isn't supported in the editor, the glyphs source format
supports curves with only on-curves, using the truetype 'implied oncurve
points' model.

When dealing with a contour of this type, we should skip rotating the
nodes when generating the path, because the first point of these
contours is already correct (being the off-curve point of the first
segment).