Work around icu4x issue with unicode 16 scriptext
This also bumps us to use the new icu4x 2.0 beta.

icu4x is not currently returning the correct script values for
codepoints that were added to ScriptExtensions.txt in Unicode 16,
so we have a temporary workaround that manually checks for those
specific values before calling out to icu4x.