[normalizer] Normalize GDEF (LigCaretList)
The LigCaretList in GDEF is another subtable that contains values that
may have associated variation data, which means that it is highly
susceptable to minor differences related to the sort order of the
item variation store.

This adds logic for generating a normalized representation of this
tables contents, simililar to how we handle marks & kerns.