[fea-rs] Tweak sort order of aalt rules
The aalt feature has special behaviour where rules are supposed to be
added to aalt from other features, in the order that those features are
listed in the aalt block.

Unfortunately, since a given feature can exist for multiple
languagesystems and contain different lookups for each, there is still
ambiguity to the order in which rules are to be added.

Previously (and without much intention) we were adding rules in the
order that the features would end up in the feature list; that is,
within a feature we were adding rules based on the sorting order of the
language systems.

This didn't match fonttools, where rules are added in the order that
they are declared in the target feature.

With this patch, we match fonttools by re-sorting the lookups for a
given feature based on the lookup id, which is an effective proxy
for "when was this rule declared".