[fea-rs] Add ability to filter ttx tests with env var
This functionality existed but only worked for the bundled fonttools
tests (for historical reasons) this expands on that so it can be used
for all the ttx-based tests, and lets it be controlled via the
FEA_FILTER_TESTS environment variable.

Also removes some redundant code, but otherwise no functional change.