[features] Share more code, do less work
In order to compile both marks & kern we need to have already compiled
the GSUB table. To avoid doing this repeatedly, we now compile it
immediately after parsing the AST, and then make it available to the
mark & kern feature writers.

We also similarly store the explicit GDEF classes, if they were in the
FEA, and try to share more code between mark & kern generation in
general.