[glyphs] Improve parsing of escape sequences
This will now handle unicode escapes, including surrogate pairs. Will
error on unpaired surrogate pairs.

In the process of this work I noticed a few other omissions related to
string parsing; the major one is that the reference impl supports the
use of single quotes to delimit strings. This doesn't seem to be in any
of our sources, but it might be worth fixing now?