[ttx_diff] Ignore lookup indices of normalized lookups
More specifically, for lookups that are generated via feature writers in
python (mark/kern). There are differences between how fontc and ufo2ft
handle generating these features, and it would add unnecessary
complexity to try and make either match the other, so I think it's
simpler to just ignore them in the diff.