[fea-rs] Replicate feature writer insert logic
This reworks how we merge external lookups to try and precisely model
the logic of the ufo2ft FeatureWriter system.

This is complicated, but it lets us avoid the possibility of having
meaningful differences in our output be the result of minor differences
in the ordering of the generated lookups.