[crater] Display repos for which we couldnt find targets
These are just displayed in a block at the top, after the summary but
before the detailed report.

They aren't included as "other failures" because those are all
associated with running particular targets, and this failure happens
before we get a target. (A target is something specific in crater, and
you can't create one wthout valid paths to the config and the source
file.)

It's kind of ugly and we might want to iterate on that, but it at least
gets the information out there.