[ttx_diff] Sort mark filter sets in GDEF
Although we now match fonttools for the assignment of _lookup_ ids, we
can differ for the ids of mark filtering sets, because these ids are
assigned when the lookup is created and are not remapped later.

The order of these ids doesn't matter, just that they are referenced
correctly in the various lookups; and that is handled by normalizer.