[glyphs] Support 'Other' subcategory
Gets at least one more GDEF matching.

Khaled mentioned elsewhere that we may want to be more permissive in
what we accept for these values, since it _can_ just be a freeform field
and we falling back to `None` might be a source of bugs, but I'm
avoiding making that larger change in favour of simple forward progress.