[crater] Always update summary after run
Currently, we decide whether or not crater should run by checking if any
of the fontc git revision, pip freeze output, or target list sha have
changed; and then after a run has finished, we will post an update if
the run's results differed at all from the previous run.

This has an annoying side effect: if we push a new commit to fontc that
doesn't cause any change observable to crater, we will never post an
update, which means we will keep doing full, useless runs (because the
git revision of the last run never changes).

This patch changes that behaviour: now if a run produces the same output
as the previous run, we will update the summary file, but we will reuse
the results file of the previous run; this saves us writing a bunch of
redundant bytes, but means the crater summary will always include an
entry for a new commit, even if nothing changed.