[crater] Update for metadata-based sources
This is a breaking change from the previous way of representing sources
which was 'repo focused', where a repo could contain multiple sources.
Now we are focused on a 'source', where it's possible that different
sources share a single repo (but have their own config files.)

This will need to be merged alongside an update to the targets.json
file in the crater repo.