[fontbe] Don't generate vf tables for static font
Previously we would always generate some of these tables and then just
skip them when we were assembling the final font; now we skip them
earlier.

In the case of gvar there is probably some additional optimization
possible, since we are still generating and storing all the empty gvar
fragments, but the plumbing there is complicated so I've left that be
for now.

Concretely this was motivated by me wanting to write a very simple test
font that would include a STAT font from FEA, and for the purposes of my
test I don't need this font to actually be variable.