[glyphs] Treat null metric as zero if overshoot exists
Glyphs does not serialize metric values if they are zero, but sometimes
metric values are _intentionally_ zero, when they would otherwise be
meaningless; we can distinguish (at least some of?) these cases because
although the metric value is zero the overshoot is (often?) a non-zero
value.

With this patch we distinguish between a metric which is missing and a
metric that exists but has no value; in this latter case we treat it as
having a value of zero.