Improve flattening nested no-export components
Currently if a no-export glyph is a component in a composite glyph,
which is also a component in another composite glyph, all the composite
glyphs will be decomposed into simple glyphs. This is because no-export
components are flattened before decomposing no-export glyphs.

Decomposing no-export glyphs first then flattening, does not decompose
the composite glyph with nested components, which keeps the file size
smaller and matches what ufo2ft does.