Pass either a real or a fake clock to workload
This is tricky because we previously stored Instants in the timer structs, and the only way to get an Instant is with a real clock! So everything has to become generic. I tried hiding the difference behind an enum, but couldn’t get it working because at some point the enum has to return different types for the different branches. So generics it is.