[crater] Rework representation of targets
This patch is kind of doing a bunch of stuff and I apologize for getting
it all tied up together.

I. Preamble

Crater, like much software, was accreted more than it was
designed. This has started to pose challanges.

The main problem is that historically we were only concerned with font
source files; repositories and config files were just a way for us to
discover sources.

Later on, we started to also care about config files, so that we could
do gftools builds; but this work brought in a bunch of assumptions about
where config files could be found, and how they related to sources.

II. This Commit

- The main purpose of this commit is to move us from being focused on
  sources to being focused on configs. We will only build something if we
  have a config, and for a given config we will only build a source that
  is named by that config.

- In addition to that, this commit tries to generally be more consistent
  about how we construct paths, and how we report errors if we cannot
  compile a given target.

- Finally, this commit adds preliminary support for 'virtual configs',
  which occur when we want to provide a config file for a repository
  that we do not otherwise want to (or are unable to) modify. These
  config files live in the google/fonts repository.

  More work will be required before we can actually build these virtual
  configs, but at least now we understand that they exist.