[crater] Support annotations
This allows users to add 'annotations' to targets, which can be used to
explain known problems and link to issues.

Annotations are added by manually editing an `annotations.json` file
stored alongside the output.

Annotations are maximally dumb: they are associated with a target and
will be visible for as long as that target produces a diff, unless they
are manually removed.