[glyphs] Match glyphsLib for curve-curve intersections
Specficially in the case where there are multiple intersections, the
curve-curve implementation does not behave the same as the line/curve or
line/line implementations.

In the latter, the first hit is always the one with the lowest `t` on
the first segment; but in the case of segment/segment intersection, the
order is based on the order in which recursion occurs.

The key here is to not sort the results, but to retain the order we get
after recursion.